# PERF: Reduce overhead of np.linalg.norm for small arrays