@seberg 2022-04-29
Thanks!  The first change, I think we might as well just do.  Since the input is an exact array (and not even a subclass).
The second, I am not a huge fan of, but if it is noticably faster for small array it seems OK.  A comment may be nice to note that it is just a micro-optimization.