@rossbar 2022-04-29
> The second, I am not a huge fan of, but if it is noticably faster for small array it seems OK. A comment may be nice to note that it is just a micro-optimization.

Agreed - are we sure that the second change actually has an appreciable impact on the benchmarks? It'd be good to know what the contribution to performance improvement is from each change.

Also re: benchmarking, we could likely get more comprehensive results with the NumPy benchmarking suite, e.g. `python runtests.py --bench-compare main bench_linalg`