@mattip 2022-05-01
> I tried running the benchmarks, but they do not work on my systems

Please open an issue with the tracebacks. This is the standard way for reporting performance enhancements in a reproducible way and allows us to track regressions. Maybe this PR improves one measure of performance but hurts others (unlikely, but stranger things have happened).