@eendebakpt 2022-05-03
I removed the specialization for float. With the ideas in https://github.com/numpy/numpy/pull/21423 (and some other ideas) the overhead from `np.sqrt` can be reduced enough that the performance advantange does not seem worth the special case.

Running benchmarks still does not work, it is due to https://github.com/numpy/numpy/issues/21288 (but it does not yet give me a solution).