@seberg 2022-05-03
Yeah, I just ran into the benchmark problem as well and opened gh-21428 for it.  I am not sure it would be super useful here anyway, because we are a bit short on tests for small arrays/scalars.

So it would be nice to add tests for the norm of very small arrays, but considering how simple it is now and that the benchmarks don't run easily, I am happy to glance over it now.

Can you clean up the `import math` and `float64`, though?