@eendebakpt 2022-05-03
Benchmark:
```
       before           after         ratio
     [fd646bd6]       [f28a2946]
     <main>           <optimize_linalg_norm>
-     16.1±0.03μs      15.3±0.08μs     0.95  bench_linalg.Linalg.time_op('norm', 'float32')
```