@seberg 2022-05-03
OK, the benchmark seems to at least notice a small difference and I don't think this adds any maintanence burden anyway in the current state.  So lets just get it in.

Thanks @eendebakpt.